Reworked main flow, added distance and angle tracking
Removed the two extra while loops and made the the robot
use just one main while loop. As part of this, we now
track whether the robot is moving or rotating and based
on this, give the command to start rotating or moving
only once (instead of every iteration of the while loop like
before).

Now also tracking each angle we turn and each straight line
distance we travel. These can eventually be used to track
where the robot is, where it started and how to get back to the
starting position.