Robot turns slightly back after rotating, stops after full circle
The robot now makes a 2 degree turn back to the white line after turning
away from it. This makes sure we don't drive away from the white line and
move more parallel to it. It does make the robot jerk around a bit more in
place.

Converted all units to mm.

Tries making a full round of the field, then stop where it started. This
is currently a bit unfinished. Tried adding bits that would automatically
detect whether we are on the defender or attacker court, so there are
values for both fields and some code that tries detecting it, but that
doesn't really work atm.