Added projects for basic PC -> NXT communication via Bluetooth #5
Basic projects for communicating with a single robot done by @kocikocev.

Added a common Server class that manages communications to two robots.
This handles connecting to the robots and sending commands to them. Right
now there is only basic support for receiving commands from the robots as
it is still under debate whether we need the robots to send info back,
for performance reasons we are likely to stick to a one-way communication
channel.