New PitchPoints class prototype to start splitting up PitchPoints. #12
This is a first version prototype. This is what we WANT the
interface to the old ObjectPoints to look like. It also provides basic support
for dividing the pitch into quadrants, each of which is a separate
ObjectPoints instance. This class will later also be responsible for
merging the information from these four quadrants.