Defender goes more in straight lines now
Defender will pick a target point, but will actually go towards either the top or bottom of the pitch (depending on the direction of the target point). This will cause it to always move at an optimal angle and to continue going without stopping if it turns out that we need to go further in the same direction.